Deal with supported input types missing from registry upon transform registration.
One input type missing from the registry is enough to make registration of the
entire transform fail. So don't try to install the transform with all the
supported input types from the start. Instead, first register it for a simple
set of initial types, then filter the list of supported types to only the ones
present in the MTR, and update the transforms input types afterwards.