Testing layers: explicitly configure standalone layer without server port.
FTW_TIKA_INTEGRATION_TESTING: The converter does no longer try to connect to
the server (which is not running) and then fallback.
We only want to test standalone conversion with this layer.

TIKA_SERVER_INTEGRATION_TESTING: The fallback to standalone is explicitly
disabled (by removing the path) and the port is configured.
This makes all conversions on this layer either use the server (which is
started here) or raise.

With this change we have no longer ERROR-log entries and we can rely on the
testing setup that it uses the conversion implementation we expect or raise
an exception.