Remove local fallback when server fails.
When the tika server requires too much memory for extracting the text
of a document, it could fail (maybe also because of a connection
timeout).
This did lead to a local fallback, causing an uncontrolled process to
consume too much memory.

It is therefore better to not have automatic fallbacks.