Drop flaky tests against deprecated plone 4.2.
As far as i've seen we no longer use plone 4.2 anywhere with ftw.tika.

And even if we do i don't really care about that old and deprecated
version any more. I think in the year 2021 we can safely assume that at
least plone 4.3 is in use for a productive deployment.